rollback xl2tpd, new version does not work
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@33989 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/xl2tpd/md5.h
@@ -4,8 +4,7 @@
 #ifdef __alpha
 typedef unsigned int uint32;
 #else
-#include <stdint.h>
-typedef uint32_t uint32;
+typedef unsigned long uint32;
 #endif
 
 struct MD5Context
